server/requirements.txt
# Web Framework
fastapi
uvicorn[standard]
python-multipart  # for file uploads

# LangChain & Ecosystem
langchain
langchain-community
langchain-core
langchain-groq

# Vectorstore
pinecone

# Embeddings
langchain-huggingface
sentence-transformers

# PDF Parsing
PyPDF  # backend for PyPDFLoader via langchain_community

# Environment Variables
python-dotenv

# Typing & Utilities
pydantic
requests
tqdm
pandas  # For data engineering pipeline

# Authentication
pymongo[bson] 
bcrypt 

# ML Experiment Tracking
mlflow
pyyaml


#  uv pip install -r requirements.txt